% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retistruct.R
\name{checkDatadir}
\alias{checkDatadir}
\title{Check the whether  directory contains valid data}
\usage{
checkDatadir(dir = NULL)
}
\arguments{
\item{dir}{Diectory to check.}
}
\value{
\code{TRUE} if \code{dir} contains valid data;
\code{FALSE} otherwise.
}
\description{
Check the whether  directory contains valid data
}
\author{
David Sterratt
}

