% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeTearRelationships}
\alias{computeTearRelationships}
\title{Compute the parent relationships for a set of tears}
\usage{
computeTearRelationships(o, V0, VB, VF)
}
\arguments{
\item{o}{\code{AnnotatedOutline} object}

\item{V0}{Apices of tears}

\item{VB}{Backward vertices of tears}

\item{VF}{Forward vertices of tears}
}
\value{
List
\item{\code{Rset}}{the set of points on the rim}
\item{\code{TFset}}{list containing indicies of points in each foward tear}
\item{\code{TBset}}{list containing indicies of points in each backward tear}
\item{\code{h}}{correspondence mapping}
\item{\code{hf}}{correspondence mapping in forward direction for
        points on boundary}
\item{\code{hb}}{correspondence mapping in backward direction for
        points on boundary}
}
\description{
Compute the parent relationships for a potential set of tears on
an \code{AnnotatedOutline}. The function throws an error if tears
overlap.
}
\author{
David Sterratt
}

