% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{Constructor for a \code{dataset} object.}
\usage{
Dataset(o, dataset, Ds, Ss, cols, raw, Gs = NULL)
}
\arguments{
\item{o}{An \code{outline} object.}

\item{dataset}{The name of the dataset}

\item{Ds}{A list of data point sets, with each set being a 2
column matrix of X and Y coordinates of data point locations. Each
item in the list should be named. Elements with these names should
also be in the \code{cols} argument (see below).}

\item{Ss}{A list of landmarks. These do not need to be named. If
any elements are  named, the names should map onto an element in
the \code{cols} argument. Any elements that are named \code{""}
will be plotted using the default colour.}

\item{cols}{A list of colours in which to plot datapoints and landmarks.}

\item{raw}{A place to put raw data in whatever format is desired.}

\item{Gs}{A list of grouped point sets, with each set being a 3
column matrix of X and Y coordinates and the value Z of the
variable at that point.  Each item in the list should be
named. Elements with these names should also be in the \code{cols}
argument.}
}
\value{
A \code{dataset} object.
}
\description{
Constructor for a \code{dataset} object.
}
\author{
David Sterratt
}

