% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{depthplot3D}
\alias{depthplot3D}
\title{Draw the "flat" outline in 3D with depth information}
\usage{
depthplot3D(r, ...)
}
\arguments{
\item{r}{\code{\link{TriangulatedOutline}} object}

\item{...}{Parameters depending on class of \code{r}}
}
\description{
Draw the "flat" outline in 3D with depth information
}
\author{
David Sterratt
}
