% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{directories}
\alias{directories}
\title{File system directories used by shinyFiles}
\format{
An object of class \code{character} of length 1.
}
\usage{
directories
}
\description{
File system directories used by shinyFiles
}
\keyword{datasets}
