\name{draw.system}
\alias{draw.system}
\title{Draw a system of lenses}
\usage{
  draw.system(S, xoff = -10, ypad = 1, add = FALSE, ...)
}
\arguments{
  \item{S}{System defined by data frame containing left
  margin of surfaces (A), radii of surfaces (R) and
  refracive indices (n).}

  \item{xoff}{x offset of left edge}

  \item{ypad}{y padding}

  \item{add}{If \code{TRUE} add to an existing plot}

  \item{...}{Graphics parameters}
}
\description{
  Draw a system of lenses
}
\author{
  David Sterratt
}

