% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\name{ensureFixedPointInRim}
\alias{ensureFixedPointInRim}
\title{Ensure that the fixed point is in the rim, not a tear}
\usage{
ensureFixedPointInRim(o)
}
\arguments{
\item{o}{\code{\link{AnnotatedOutline}} object}
}
\value{
o \code{\link{AnnotatedOutline}} object in which \code{i0}
may have been changed.
}
\description{
Ensure that the fixed point \code{i0} is in the rim, not a tear.
}
\author{
David Sterratt
}

