% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spheristruct.R
\name{f}
\alias{f}
\title{Piecewise smooth function used in area penalty}
\usage{
f(x, x0)
}
\arguments{
\item{x}{Main argument}

\item{x0}{The cutoff parameter. Above this value the function is zero.}
}
\value{
The value of the function.
}
\description{
Piecewise, smooth function that increases linearly with negative arguments.
\deqn{    f(x) = \left\{
       \begin{array}{ll}
         -(x - x_0/2) & x < 0 \\
         \frac{1}{2x_0}(x - x_0)^2 & 0 < x <x_0 \\
         0 & x \ge x_0
         \end{array} \right.
}
}
\author{
David Sterratt
}

