% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSet.R
\name{FeatureSet}
\alias{FeatureSet}
\title{FeatureSet class}
\value{
An \code{FeatureSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
FeatureSet class

FeatureSet class
}
\author{
David Sterratt
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureSet$new()}}
\item \href{#method-getID}{\code{FeatureSet$getID()}}
\item \href{#method-getIDs}{\code{FeatureSet$getIDs()}}
\item \href{#method-setName}{\code{FeatureSet$setName()}}
\item \href{#method-getFeature}{\code{FeatureSet$getFeature()}}
\item \href{#method-getFeatures}{\code{FeatureSet$getFeatures()}}
\item \href{#method-getCol}{\code{FeatureSet$getCol()}}
\item \href{#method-clone}{\code{FeatureSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$new(data = NULL, cols = NULL, type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\subsection{Method \code{getID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$getID(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIDs"></a>}}
\subsection{Method \code{getIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$getIDs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setName"></a>}}
\subsection{Method \code{setName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$setName(i, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeature"></a>}}
\subsection{Method \code{getFeature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$getFeature(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeatures"></a>}}
\subsection{Method \code{getFeatures()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$getFeatures()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCol"></a>}}
\subsection{Method \code{getCol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$getCol(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
