% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\name{flatplot.AnnotatedOutline}
\alias{flatplot.AnnotatedOutline}
\title{Flat plot of AnnotatedOutline}
\usage{
\method{flatplot}{AnnotatedOutline}(x, axt = "n", xlim = NULL,
  ylim = NULL, markup = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{AnnotatedOutline}} object}

\item{axt}{whether to plot axes}

\item{xlim}{x-limits}

\item{ylim}{y-limits}

\item{markup}{If \code{TRUE}, plot markup}

\item{...}{Other plotting parameters}
}
\description{
Plot flat \code{\link{AnnotatedOutline}}. The user markup is
displayed by default.
}
\author{
David Sterratt
}
