% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outline.R
\name{flatplot.Outline}
\alias{flatplot.Outline}
\title{Flat plot of outline}
\usage{
\method{flatplot}{Outline}(
  x,
  axt = "n",
  xlim = NULL,
  ylim = NULL,
  add = FALSE,
  image = TRUE,
  scalebar = 1,
  rimset = FALSE,
  pids = FALSE,
  pid.joggle = 0,
  lwd.outline = 1,
  ...
)
}
\arguments{
\item{x}{\code{\link{Outline}} object}

\item{axt}{whether to plot axes}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{add}{If \code{TRUE}, don't draw axes; add to existing plot.}

\item{image}{If \code{TRUE} the image (if it is present) is
displayed behind the outline}

\item{scalebar}{If  numeric and if the Outline has a \code{scale}
field, a scale bar of length \code{scalebar} mm is plotted.  If
\code{scalebar} is \code{FALSE} or there is no scale information
in the \code{\link{Outline}} \code{x} the scale bar is suppressed.}

\item{rimset}{If \code{TRUE}, plot the points computed to be in the rim in the colour specified by the option \code{rimset.col}}

\item{pids}{If \code{TRUE}, plot point IDs}

\item{pid.joggle}{Amount to joggle point IDs by randomly}

\item{lwd.outline}{Line width of outline}

\item{...}{Other plotting parameters}
}
\description{
Plot flat \code{\link{Outline}}.
}
\author{
David Sterratt
}
