% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedOutline.R
\name{flatplot.ReconstructedOutline}
\alias{flatplot.ReconstructedOutline}
\title{Flat plot of reconstructed outline}
\usage{
\method{flatplot}{ReconstructedOutline}(
  x,
  axt = "n",
  xlim = NULL,
  ylim = NULL,
  grid = TRUE,
  strain = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{\link{ReconstructedOutline}} object}

\item{axt}{whether to plot axes}

\item{xlim}{x-limits}

\item{ylim}{y-limits}

\item{grid}{Whether or not to show the grid lines of
latitude and longitude}

\item{strain}{Whether or not to show the strain}

\item{...}{Other plotting parameters}
}
\description{
Plot \code{\link{ReconstructedOutline}} object. This adds a mesh
of gridlines from the spherical retina (described by points
\code{phi}, \code{lambda} and triangulation \code{Trt} and cut-off
point \code{phi0}) onto a flattened retina (described by points
\code{P} and triangulation \code{T}).
}
\author{
David Sterratt
}
