% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangulatedOutline.R
\name{flatplot.TriangulatedOutline}
\alias{flatplot.TriangulatedOutline}
\title{Plot flat \code{\link{TriangulatedOutline}}.}
\usage{
\method{flatplot}{TriangulatedOutline}(x, axt = "n", xlim = NULL, ylim = NULL, mesh = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{TriangulatedOutline}} object}

\item{axt}{whether to plot axes}

\item{xlim}{x-limits}

\item{ylim}{y-limits}

\item{mesh}{If \code{TRUE}, plot mesh}

\item{...}{Other plotting parameters}
}
\description{
Plot flat \code{\link{TriangulatedOutline}}.
}
\author{
David Sterratt
}
