% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fragment.R
\docType{data}
\name{Fragment}
\alias{Fragment}
\title{Fragment constructor}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Fragment
}
\value{
An \code{Fragment} object containing the following:
\item{\code{P}}{A N-by-2 matrix of points of the \code{Outline} arranged in anticlockwise order}
\item{\code{gf}}{For each row of \code{P}, the index of \code{P} that is next in the outline travelling anticlockwise (forwards)}
\item{\code{gb}}{For each row of \code{P}, the index of \code{P} that is next in the outline travelling clockwise (backwards)}
\item{\code{h}}{For each row of \code{P}, the correspondence of that point (which will be to itself initially)}
}
\description{
Construct an outline object. This sanitises the input points
\code{P}, as described below.
}
\author{
David Sterratt
}
\keyword{datasets}
