% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReconstructedDataset.R
\name{getDss.reconstructedDataset}
\alias{getDss.reconstructedDataset}
\title{Get transformed spherical coordinates of datapoints}
\usage{
\method{getDss}{reconstructedDataset}(r)
}
\arguments{
\item{r}{\code{ReconstructedDataset} object.}
}
\value{
\code{Dss}
}
\description{
Get spherical coordinates of datapoints.
}
\author{
David Sterratt
}

