% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedDataset.R
\name{getDssHullarea}
\alias{getDssHullarea}
\title{Get area of convex hull around data points on sphere}
\usage{
getDssHullarea(r)
}
\arguments{
\item{r}{code{\link{ReconstructedDataset}} or \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
Area in degrees squared
}
\description{
Get area of convex hull around data points on sphere
}
\author{
David Sterratt
}
