% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{getDssMean}
\alias{getDssMean}
\title{Karcher mean of datapoints in spherical coordinates}
\usage{
getDssMean(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} or \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Dss.mean}
}
\description{
Get Karcher mean of datapoints in spherical coordinates.
}
\author{
David Sterratt
}

