% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReconstructedDataset.R
\name{getDssMean.reconstructedDataset}
\alias{getDssMean.reconstructedDataset}
\title{Karcher mean of datapoints in spherical coordinates}
\usage{
\method{getDssMean}{reconstructedDataset}(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} or \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Dss.mean}
}
\description{
Get Karcher mean of datapoints in spherical coordinates.
}
\author{
David Sterratt
}

