% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RetinalReconstructedDataset.R
\name{getDssMean.retinalReconstructedDataset}
\alias{getDssMean.retinalReconstructedDataset}
\title{Get transformed spherical coordinates of Karcher mean of
datapoints}
\usage{
\method{getDssMean}{retinalReconstructedDataset}(r)
}
\arguments{
\item{r}{\code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Dss.mean}
}
\description{
Get Karcher mean of datapoints in spherical coordinates,
transformed according to the values of \code{DVflip} and
\code{side}.
}
\author{
David Sterratt
}

