% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\name{getFlatRimLength}
\alias{getFlatRimLength}
\title{Get rim length of AnnotatedOutline}
\usage{
getFlatRimLength(o)
}
\arguments{
\item{o}{\code{\link{AnnotatedOutline}} object}
}
\value{
The rim length
}
\description{
Get rim length of AnnotatedOutline
}
\author{
David Sterratt
}

