% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{getGss}
\alias{getGss}
\title{Get grouped variable with locations in spherical coordinates.}
\usage{
getGss(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} or \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Gss}
}
\description{
Get grouped variable with locations in spherical coordinates.
}
\author{
David Sterratt
}

