% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGss.retinalReconstructedDataset}
\alias{getGss.retinalReconstructedDataset}
\title{Get grouped variable with locations in spherical coordinates.}
\usage{
\method{getGss}{retinalReconstructedDataset}(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} or \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Gss}
}
\description{
Get grouped variable with locations in spherical coordinates.
}
\author{
David Sterratt
}

