% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Dataset.R
\name{getIDs.dataset}
\alias{getIDs.dataset}
\title{Get IDs of groups of data within a dataset}
\usage{
\method{getIDs}{dataset}(r)
}
\arguments{
\item{r}{\code{\link{Dataset}} object}
}
\value{
Array of IDs
}
\description{
Get IDs of groups of data within a dataset
}
\author{
David Sterratt
}

