% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReconstructedDataset.R
\name{getIDs.reconstructedDataset}
\alias{getIDs.reconstructedDataset}
\title{Get IDs of groups of data within a ReconstructedDataset}
\usage{
\method{getIDs}{reconstructedDataset}(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} object}
}
\value{
Array of IDs
}
\description{
Get IDs of groups of data within a ReconstructedDataset
}
\author{
David Sterratt
}

