% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getIms}
\alias{getIms}
\title{Get coordinates of corners of pixels of image in spherical coordinates}
\usage{
getIms(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedOutline}} or
\code{\link{RetinalReconstructedOutline}} object}
}
\value{
Coordinates of corners of pixels in spherical coordinates
}
\description{
Get coordinates of corners of pixels of image in spherical coordinates
}
\author{
David Sterratt
}

