% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReconstructedOutline.R
\name{getIms.reconstructedOutline}
\alias{getIms.reconstructedOutline}
\title{Get coordinates of corners of pixels of image in spherical
coordinates}
\usage{
\method{getIms}{reconstructedOutline}(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedOutline}} object}
}
\value{
Coordinates of corners of pixels in spherical coordinates
}
\description{
Get coordinates of corners of pixels of image in spherical
coordinates
}
\author{
David Sterratt
}

