% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{getSss}
\alias{getSss}
\title{Get transformed spherical coordinates of landmarks.}
\usage{
getSss(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} or \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Sss}
}
\description{
Get spherical coordinates of landmarks.
}
\author{
David Sterratt
}

