% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RetinalReconstructedDataset.R
\name{getSss.retinalReconstructedDataset}
\alias{getSss.retinalReconstructedDataset}
\title{Get transformed spherical coordinates of datapoints}
\usage{
\method{getSss}{retinalReconstructedDataset}(r)
}
\arguments{
\item{r}{\code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Dss}
}
\description{
Get spherical coordinates of landmarks, transformed according to
the values of \code{DVflip} and \code{side}.
}
\author{
David Sterratt
}

