% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSssMean}
\alias{getSssMean}
\title{Karcher mean of landmarks in spherical coordinates}
\usage{
getSssMean(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} or \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
List of spherical coordinates, with one named element per
landmark.
}
\description{
Get Karcher mean of landmarks in spherical coordinates. If applied
to a \code{\link{RetinalReconstructedDataset}}, the coordinates
are flipped if \code{DVflip} is \code{TRUE}.
}
\author{
David Sterratt
}

