% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\name{getTear}
\alias{getTear}
\title{Return indices of tear in AnnotatedOutline}
\usage{
getTear(o, tid)
}
\arguments{
\item{o}{\code{AnnotatedOutline} object}

\item{tid}{Tear ID, which can be returned from \code{whichTear()}}
}
\value{
Vector of three point IDs, labelled with \code{V0},
\code{VF} and \code{VB}
}
\description{
Return indices of tear in AnnotatedOutline
}
\author{
David Sterratt
}
