\name{getTss.reconstructedOutline}
\alias{getTss.reconstructedOutline}
\title{Get spherical coordinates of tears.}
\usage{
  \method{getTss}{reconstructedOutline} (r)
}
\arguments{
  \item{r}{\code{reconstructedOutline} object.}
}
\value{
  \code{Tss}
}
\description{
  Get spherical coordinates of tears.
}
\author{
  David Sterratt
}

