\name{getTss.retinalReconstructedOutline}
\alias{getTss.retinalReconstructedOutline}
\title{Get spherical coordinates of tears.}
\usage{
  \method{getTss}{retinalReconstructedOutline} (r)
}
\arguments{
  \item{r}{\code{\link{RetinalReconstructedOutline}}
  object.}
}
\value{
  List of tears in spherical coordinates.
}
\description{
  This flips the tears according to whether the retina is
  DV flipped or not.
}
\author{
  David Sterratt
}

