% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformations.R
\name{identity.transform}
\alias{identity.transform}
\title{The identity transformation}
\usage{
identity.transform(r, ...)
}
\arguments{
\item{r}{Coordinates of points in spherical coordinates
represented as  2 column matrix with column names \code{phi}
(lattitude) and \code{lambda} (longitude).}

\item{...}{Other arguments}
}
\value{
Idendtical matrix
}
\description{
The identity transformation
}
\author{
David Sterratt
}

