% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-ijroi-multi.R
\name{ijroimulti.read.dataset}
\alias{ijroimulti.read.dataset}
\title{Read a retinal dataset in IJROI format}
\usage{
ijroimulti.read.dataset(dataset)
}
\arguments{
\item{dataset}{Path to directory containing \code{outline.roi}}
}
\value{
A \code{\link{RetinalOutline}} object
}
\description{
Read a retinal dataset in IJROI format. Each dataset is a folder
containing a file called \code{outline.roi} that specifies the
outline in X-Y coordinates. It may also contain a file
\code{datapoints.csv}, containing the locations of data points;
see \code{\link{read.datapoints}} for the format of this file. The
folder may also contain a file \code{od.roi} specifying the
coordinates of the optic disc.
}
\author{
David Sterratt
}
