\name{kde.compute.concentration}
\alias{kde.compute.concentration}
\title{Find the optimal concentration for a set of data}
\usage{
kde.compute.concentration(mu)
}
\arguments{
  \item{mu}{Data in spherical coordinates}
}
\value{
The optimal concentration
}
\description{
Find the optimal concentration for a set of data
}
\author{
David Sterratt
}

