\name{kde.fhat}
\alias{kde.fhat}
\title{Kernel density estimate on sphere using Fisherian density
with polar coordinates}
\usage{
  kde.fhat(r, mu, kappa)
}
\arguments{
  \item{r}{Locations at which to estimate density in polar
  coordinates}

  \item{mu}{Locations of data points in polar coordinates}

  \item{kappa}{Concentration parameter}
}
\value{
  Vector of density estimates
}
\description{
  Kernel density estimate on sphere using Fisherian density
  with polar coordinates
}
\author{
  David Sterratt
}

