% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6-serialize.R
\name{list_to_R6}
\alias{list_to_R6}
\title{Convert an list created by R6_to_list() into an R6 object.}
\usage{
list_to_R6(l)
}
\arguments{
\item{l}{list created by R6_to_list()}
}
\value{
R6 object or list list
}
\description{
Convert an list created by R6_to_list() into an R6 object.
}
\author{
David Sterratt
}
