% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{lvsLplot}
\alias{lvsLplot}
\title{Plot the fractional change in length of mesh edges}
\usage{
lvsLplot(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedOutline}} object}
}
\description{
Plot the fractional change in length of mesh edges. The length of
each edge in the mesh in the reconstructed object is plotted
against each edge in the spherical object. The points are
colour-coded according to the amount of log strain each edge is
under.
}
\author{
David Sterratt
}

