\name{lvsLplot.reconstructedOutline}
\alias{lvsLplot.reconstructedOutline}
\title{Plot the fractional change in length of mesh edges}
\usage{
  \method{lvsLplot}{reconstructedOutline} (r)
}
\arguments{
  \item{r}{\code{reconstructedOutline} object}
}
\description{
  Plot the fractional change in length of mesh edges. The
  length of each edge in the mesh in the reconstructed
  object is plotted against each edge in the spherical
  object. The points are colour-coded according to the
  amount of log strain each edge is under.
}
\author{
  David Sterratt
}

