% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Magnifier.R
\name{Magnifier}
\alias{Magnifier}
\title{Class to allow ggraphics objects to be magnified and scrolled}
\description{
This creates a wrapper round
  \code{\link[gWidgets2]{ggraphics}} that adds + and - buttons to
  allow the user to resize the device. If the device is bigger
  than the plot window, then scrollbars appear.
}
\examples{
\dontrun{
w <- gWidgets2::gwindow("Magnifier", height=400, width=400)
# Create the mangfier instance
m  <- Magnifier$new(container=w, ps=11)
## Add extra buttons
g.print     <- gWidgets2::gbutton("Bitmap", container=m$buttons)
## Plot in the window
plot(1:10, 1:10)
## Set the device to be the magnifier
m$devSet()
plot(1:5, -5:-1)
## Access the device with m$d
dev.set(m$d)
}
}
\author{
David Sterratt
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{d}}{Device handle}

\item{\code{buttons}}{gWidgets2::ggroup handler providing space for buttons}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Magnfier-new}{\code{Magnifier$new()}}
\item \href{#method-Magnfier-devSet}{\code{Magnifier$devSet()}}
\item \href{#method-Magnfier-clone}{\code{Magnifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Magnfier-new"></a>}}
\if{latex}{\out{\hypertarget{method-Magnfier-new}{}}}
\subsection{Method \code{new()}}{
Magnifier constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magnifier$new(container, width = 500, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container}}{gWidgets2 parent container}

\item{\code{width}}{The width of the Magnifier in pixels}

\item{\code{...}}{Arguments passed to \code{\link[gWidgets2]{ggraphics}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Magnfier-devSet"></a>}}
\if{latex}{\out{\hypertarget{method-Magnfier-devSet}{}}}
\subsection{Method \code{devSet()}}{
Set the graphics device to the device contained in the
magnifier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magnifier$devSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Magnfier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Magnfier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magnifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
