% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tests.R
\name{morph.dataset.to.parabola}
\alias{morph.dataset.to.parabola}
\title{Morph a flat dataset to a parabola for testing purposes}
\usage{
morph.dataset.to.parabola(
  orig.dataset = file.path(system.file(package = "retistruct"), "extdata", "smi32-csv"),
  morphed.dataset = NA,
  f = 100
)
}
\arguments{
\item{orig.dataset}{Directory of original dataset}

\item{morphed.dataset}{Directory to write morphed dataset to. If
NA, a temporary directory will be created}

\item{f}{Focus of parabola}
}
\value{
Path to \code{morphed.dataset}
}
\description{
Morph a flat dataset to a parabola for testing purposes
}
\author{
David Sterratt
}
