% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{name.list}
\alias{name.list}
\title{Return a new version of the list in which any unnamed elements
have been given standardised names}
\usage{
name.list(l)
}
\arguments{
\item{l}{the list with unnamed elements}
}
\value{
The list with standardised names
}
\description{
Return a new version of the list in which any unnamed elements
have been given standardised names
}
\author{
David Sterratt
}
