% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{nameLandmark}
\alias{nameLandmark}
\title{Name a landmark in a Dataset}
\usage{
nameLandmark(d, i, name)
}
\arguments{
\item{d}{\code{\link{Dataset}} object}

\item{i}{index of landmark to name}

\item{name}{name to give landmark}
}
\value{
New \code{\link{Dataset}} object in which landmark is
named
}
\description{
Name a landmark in a \code{\link{Dataset}}. The name of element
\code{i} of \code{Ss} is set to \code{name}, the name of any
element that bore the name is set to "" and all other elements are
unaltered.
}
\author{
David Sterratt
}

