% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nameLandmark.retinalDataset}
\alias{nameLandmark.retinalDataset}
\title{Name a landmark in a RetinalDataset}
\usage{
\method{nameLandmark}{retinalDataset}(d, i, name)
}
\arguments{
\item{d}{\code{\link{RetinalDataset}} object}

\item{i}{index of landmark to name}

\item{name}{name to give landmark}
}
\value{
New \code{\link{RetinalDataset}} object in which landmark
is named
}
\description{
Name a landmark in a \code{\link{RetinalDataset}}. This does the
same as the standard \code{\link{nameLandmark}}, but in addition,
if there exists a landmark named "OD", this creates a set of
points labelled "OD".
}
\author{
David Sterratt
}

