% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometry.R
\name{normalise.angle}
\alias{normalise.angle}
\title{Bring angle into range}
\usage{
normalise.angle(theta)
}
\arguments{
\item{theta}{Angle to bring into range \code{[-pi, pi]}}
}
\value{
Normalised angle
}
\description{
Bring angle into range
}
\author{
David Sterratt
}

