% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{orthographic}
\alias{orthographic}
\title{Orthographic projection}
\usage{
orthographic(r, proj.centre = cbind(phi = 0, lambda = 0), ...)
}
\arguments{
\item{r}{Latitude-longitude coordinates in a matrix with columns
labelled \code{phi} (latitude) and \code{lambda} (longitude)}

\item{proj.centre}{Location of centre of projection as matrix with
column names \code{phi} (elevation) and \code{lambda} (longitude).}

\item{...}{Arguments not used by this projection.n}
}
\value{
Two-column matrix with columns labelled \code{x} and
\code{y} of locations of projection of coordinates on plane
}
\description{
Orthographic projection
}
\references{
\url{https://en.wikipedia.org/wiki/Map_projection},
\url{http://mathworld.wolfram.com/OrthographicProjection.html}
}
\author{
David Sterratt
}
