% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outline.R
\name{Outline}
\alias{Outline}
\title{Outline class}
\value{
An \code{Outline} object containing the following:
\item{\code{P}}{A N-by-2 matrix of points of the \code{Outline} arranged in anticlockwise order}
\item{\code{gf}}{For each row of \code{P}, the index of \code{P} that is next in the outline travelling anticlockwise (forwards)}
\item{\code{gb}}{For each row of \code{P}, the index of \code{P} that is next in the outline travelling clockwise (backwards)}
\item{\code{h}}{For each row of \code{P}, the correspondence of that point (which will be to itself initially)}
\item{\code{im}}{The image as a \code{raster} object}
\item{\code{scale}}{The length of one unit of \code{P} in arbitrary units}
}
\description{
Construct an outline object. This sanitises the input points
\code{P}, as described below.
}
\author{
David Sterratt
}
\section{Super class}{
\code{\link[retistruct:OutlineCommon]{retistruct::OutlineCommon}} -> \code{Outline}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Outline$new()}}
\item \href{#method-getImage}{\code{Outline$getImage()}}
\item \href{#method-replaceImage}{\code{Outline$replaceImage()}}
\item \href{#method-mapFragment}{\code{Outline$mapFragment()}}
\item \href{#method-mapPids}{\code{Outline$mapPids()}}
\item \href{#method-addPoints}{\code{Outline$addPoints()}}
\item \href{#method-getPoints}{\code{Outline$getPoints()}}
\item \href{#method-getPointsScaled}{\code{Outline$getPointsScaled()}}
\item \href{#method-getRimSet}{\code{Outline$getRimSet()}}
\item \href{#method-getOutlineSet}{\code{Outline$getOutlineSet()}}
\item \href{#method-getOutlineLengths}{\code{Outline$getOutlineLengths()}}
\item \href{#method-addFeatureSet}{\code{Outline$addFeatureSet()}}
\item \href{#method-clone}{\code{Outline$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="clearFeatureSets">}\href{../../retistruct/html/OutlineCommon.html#method-clearFeatureSets}{\code{retistruct::OutlineCommon$clearFeatureSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="getFeatureSet">}\href{../../retistruct/html/OutlineCommon.html#method-getFeatureSet}{\code{retistruct::OutlineCommon$getFeatureSet()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="getFeatureSets">}\href{../../retistruct/html/OutlineCommon.html#method-getFeatureSets}{\code{retistruct::OutlineCommon$getFeatureSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="getFeatureSetTypes">}\href{../../retistruct/html/OutlineCommon.html#method-getFeatureSetTypes}{\code{retistruct::OutlineCommon$getFeatureSetTypes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="getIDs">}\href{../../retistruct/html/OutlineCommon.html#method-getIDs}{\code{retistruct::OutlineCommon$getIDs()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$new(P = NULL, scale = NA, im = NULL, units = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getImage"></a>}}
\subsection{Method \code{getImage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$getImage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replaceImage"></a>}}
\subsection{Method \code{replaceImage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$replaceImage(im)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapFragment"></a>}}
\subsection{Method \code{mapFragment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$mapFragment(fragment, pids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapPids"></a>}}
\subsection{Method \code{mapPids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$mapPids(x, y, pids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addPoints"></a>}}
\subsection{Method \code{addPoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$addPoints(P)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPoints"></a>}}
\subsection{Method \code{getPoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$getPoints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPointsScaled"></a>}}
\subsection{Method \code{getPointsScaled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$getPointsScaled()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRimSet"></a>}}
\subsection{Method \code{getRimSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$getRimSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOutlineSet"></a>}}
\subsection{Method \code{getOutlineSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$getOutlineSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOutlineLengths"></a>}}
\subsection{Method \code{getOutlineLengths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$getOutlineLengths()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addFeatureSet"></a>}}
\subsection{Method \code{addFeatureSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$addFeatureSet(fs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
