% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{parabola.arclength}
\alias{parabola.arclength}
\title{Arc length of a parabola y=x^2/4f}
\usage{
parabola.arclength(x1, x2, f)
}
\arguments{
\item{x1}{x co-ordinate of start of arc}

\item{x2}{x co-ordinate of end of arc}

\item{f}{focal length of parabola}
}
\value{
length of parabola arc
}
\description{
Arc length of a parabola y=x^2/4f
}
\author{
David Sterratt
}
