% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{parabola.invarclength}
\alias{parabola.invarclength}
\title{Inverse arc length of a parabola y=x^2/4f}
\usage{
parabola.invarclength(x1, s, f)
}
\arguments{
\item{x1}{co-ordinate of start of arc}

\item{s}{length of parabola arc to follow}

\item{f}{focal length of parabola}
}
\value{
x co-ordinate of end of arc
}
\description{
Inverse arc length of a parabola y=x^2/4f
}
\author{
David Sterratt
}
