% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse.dependencies}
\alias{parse.dependencies}
\title{Parse dependencies}
\usage{
parse.dependencies(deps)
}
\arguments{
\item{deps}{Text produced by, e.g., installed.packages()["packagename","Suggests"]}
}
\value{
Table with package column, relationship column and version number
}
\description{
Parse dependencies
}
\author{
David Sterratt
}
