% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSet.R
\name{PointSet}
\alias{PointSet}
\title{PointSet class}
\value{
An \code{PointSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
PointSet class

PointSet class
}
\author{
David Sterratt
}
\section{Super class}{
\code{\link[retistruct:FeatureSet]{retistruct::FeatureSet}} -> \code{PointSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PointSet$new()}}
\item \href{#method-reconstruct}{\code{PointSet$reconstruct()}}
\item \href{#method-clone}{\code{PointSet$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSet" data-id="getCol">}\href{../../retistruct/html/FeatureSet.html#method-getCol}{\code{retistruct::FeatureSet$getCol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSet" data-id="getFeature">}\href{../../retistruct/html/FeatureSet.html#method-getFeature}{\code{retistruct::FeatureSet$getFeature()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSet" data-id="getFeatures">}\href{../../retistruct/html/FeatureSet.html#method-getFeatures}{\code{retistruct::FeatureSet$getFeatures()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSet" data-id="getID">}\href{../../retistruct/html/FeatureSet.html#method-getID}{\code{retistruct::FeatureSet$getID()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSet" data-id="getIDs">}\href{../../retistruct/html/FeatureSet.html#method-getIDs}{\code{retistruct::FeatureSet$getIDs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSet" data-id="setName">}\href{../../retistruct/html/FeatureSet.html#method-setName}{\code{retistruct::FeatureSet$setName()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointSet$new(data = NULL, cols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reconstruct"></a>}}
\subsection{Method \code{reconstruct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointSet$reconstruct(ro)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
