% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{polartext}
\alias{polartext}
\title{Put text on the polar plot}
\usage{
polartext(text)
}
\arguments{
\item{text}{Test to place}
}
\description{
Place text at bottom right of \code{\link{projection}}
}
\author{
David Sterratt
}

