% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{projection}
\alias{projection}
\title{Plot projection of a reconstructed outline}
\usage{
projection(r, ...)
}
\arguments{
\item{r}{Object such as a \code{\link{ReconstructedOutline}}}

\item{...}{Other plotting parameters}
}
\description{
Plot projection of a reconstructed outline
}
\author{
David Sterratt
}
