% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6-serialize.R
\name{R6_to_list}
\alias{R6_to_list}
\title{Convert an R6 object into a list, ignoring functions and
environments}
\usage{
R6_to_list(r, path = "", envs = list())
}
\arguments{
\item{r}{R6 object or list}

\item{path}{root of the path to the list - no need to supply. Not
used but could be developed for pretty-printing}

\item{envs}{list of environments already encountered - do not set}
}
\value{
List with structure mirroring the R6 object.
}
\description{
Convert an R6 object into a list, ignoring functions and
environments
}
\author{
David Sterratt
}
