% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct.R
\docType{data}
\name{recfile.version}
\alias{recfile.version}
\title{Version of reconstruction file data format}
\format{An object of class \code{numeric} of length 1.}
\usage{
recfile.version
}
\description{
Version of reconstruction file data format
}
\keyword{datasets}
