% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedCountSet.R
\name{ReconstructedCountSet}
\alias{ReconstructedCountSet}
\title{ReconstructedCountSet class}
\value{
An \code{ReconstructedCountSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
ReconstructedCountSet class

ReconstructedCountSet class
}
\author{
David Sterratt
}
\section{Super class}{
\code{\link[retistruct:ReconstructedFeatureSet]{retistruct::ReconstructedFeatureSet}} -> \code{ReconstructedCountSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReconstructedCountSet$new()}}
\item \href{#method-getKR}{\code{ReconstructedCountSet$getKR()}}
\item \href{#method-clone}{\code{ReconstructedCountSet$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getCol">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getCol}{\code{retistruct::ReconstructedFeatureSet$getCol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getFeature">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getFeature}{\code{retistruct::ReconstructedFeatureSet$getFeature()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getID">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getID}{\code{retistruct::ReconstructedFeatureSet$getID()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getIDs">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getIDs}{\code{retistruct::ReconstructedFeatureSet$getIDs()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedCountSet$new(fs = NULL, ro = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getKR"></a>}}
\subsection{Method \code{getKR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedCountSet$getKR()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedCountSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
