% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedFeatureSet.R
\name{ReconstructedFeatureSet}
\alias{ReconstructedFeatureSet}
\title{ReconstructedFeatureSet class
Initialised with a FeatureSet (\code{fs}) and an ReconstructedOutline (\code{ro}) object}
\value{
A \code{ReconstructedFeatureSet} object.
}
\description{
ReconstructedFeatureSet class
Initialised with a FeatureSet (\code{fs}) and an ReconstructedOutline (\code{ro}) object

ReconstructedFeatureSet class
Initialised with a FeatureSet (\code{fs}) and an ReconstructedOutline (\code{ro}) object
}
\author{
David Sterratt
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReconstructedFeatureSet$new()}}
\item \href{#method-getID}{\code{ReconstructedFeatureSet$getID()}}
\item \href{#method-getIDs}{\code{ReconstructedFeatureSet$getIDs()}}
\item \href{#method-getFeature}{\code{ReconstructedFeatureSet$getFeature()}}
\item \href{#method-getCol}{\code{ReconstructedFeatureSet$getCol()}}
\item \href{#method-clone}{\code{ReconstructedFeatureSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedFeatureSet$new(fs = NULL, ro = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\subsection{Method \code{getID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedFeatureSet$getID(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIDs"></a>}}
\subsection{Method \code{getIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedFeatureSet$getIDs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeature"></a>}}
\subsection{Method \code{getFeature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedFeatureSet$getFeature(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCol"></a>}}
\subsection{Method \code{getCol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedFeatureSet$getCol(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedFeatureSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
