% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedLandmarkSet.R
\docType{data}
\name{ReconstructedLandmarkSet}
\alias{ReconstructedLandmarkSet}
\title{ReconstructedLandmarkSet class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReconstructedLandmarkSet
}
\value{
An \code{ReconstructedLandmarkSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
ReconstructedLandmarkSet class
}
\author{
David Sterratt
}
\keyword{datasets}
