% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedOutline.R
\name{ReconstructedOutline}
\alias{ReconstructedOutline}
\title{Reconstruct outline into spherical surface}
\description{
Reconstruct outline into spherical surface. Reconstruction
proceeds in a number of stages:
}
\details{
\enumerate{

\item The flat object is triangulated with at least \code{n}
triangles. This can introduce new vertices in the rim. 

\item The triangulated object is stitched.

\item The stitched object is triangulated again, but this time it
is not permitted to add extra vertices to the rim.

\item The corresponding points determined by the stitching process
are merged to form a new set of merged points and a new
triangulation.

\item The merged points are projected roughly to a sphere.

\item The locations of the points on the sphere are moved so as to
minimise the energy function.
}
}
\section{Member functions}{


\describe{
\item{mapFlatToSpherical(P)}{Returns location of point on sphere corresponding to point on the flat outline. Input values:}
\describe{
\item{P}{Cartesian coordinates  on flat outline as a matrix with "X" and "Y" columns}
}}
}

\author{
David Sterratt
}
\section{Super class}{
\code{\link[retistruct:OutlineCommon]{retistruct::OutlineCommon}} -> \code{ReconstructedOutline}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-loadOutline}{\code{ReconstructedOutline$loadOutline()}}
\item \href{#method-reconstruct}{\code{ReconstructedOutline$reconstruct()}}
\item \href{#method-mergePointsEdges}{\code{ReconstructedOutline$mergePointsEdges()}}
\item \href{#method-projectToSphere}{\code{ReconstructedOutline$projectToSphere()}}
\item \href{#method-getStrains}{\code{ReconstructedOutline$getStrains()}}
\item \href{#method-optimiseMapping}{\code{ReconstructedOutline$optimiseMapping()}}
\item \href{#method-optimiseMappingCart}{\code{ReconstructedOutline$optimiseMappingCart()}}
\item \href{#method-transformImage}{\code{ReconstructedOutline$transformImage()}}
\item \href{#method-getIms}{\code{ReconstructedOutline$getIms()}}
\item \href{#method-getTearCoords}{\code{ReconstructedOutline$getTearCoords()}}
\item \href{#method-getFeatureSet}{\code{ReconstructedOutline$getFeatureSet()}}
\item \href{#method-reconstructFeatureSets}{\code{ReconstructedOutline$reconstructFeatureSets()}}
\item \href{#method-getPoints}{\code{ReconstructedOutline$getPoints()}}
\item \href{#method-mapFlatToSpherical}{\code{ReconstructedOutline$mapFlatToSpherical()}}
\item \href{#method-clone}{\code{ReconstructedOutline$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="clearFeatureSets">}\href{../../retistruct/html/OutlineCommon.html#method-clearFeatureSets}{\code{retistruct::OutlineCommon$clearFeatureSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="getFeatureSets">}\href{../../retistruct/html/OutlineCommon.html#method-getFeatureSets}{\code{retistruct::OutlineCommon$getFeatureSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="getFeatureSetTypes">}\href{../../retistruct/html/OutlineCommon.html#method-getFeatureSetTypes}{\code{retistruct::OutlineCommon$getFeatureSetTypes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="OutlineCommon" data-id="getIDs">}\href{../../retistruct/html/OutlineCommon.html#method-getIDs}{\code{retistruct::OutlineCommon$getIDs()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadOutline"></a>}}
\subsection{Method \code{loadOutline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$loadOutline(
  ol,
  n = 500,
  alpha = 8,
  x0 = 0.5,
  plot.3d = FALSE,
  dev.flat = NA,
  dev.polar = NA,
  report = retistruct::report,
  debug = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reconstruct"></a>}}
\subsection{Method \code{reconstruct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$reconstruct(
  plot.3d = FALSE,
  dev.flat = NA,
  dev.polar = NA,
  report = getOption("retistruct.report")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergePointsEdges"></a>}}
\subsection{Method \code{mergePointsEdges()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$mergePointsEdges()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projectToSphere"></a>}}
\subsection{Method \code{projectToSphere()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$projectToSphere()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStrains"></a>}}
\subsection{Method \code{getStrains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$getStrains()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimiseMapping"></a>}}
\subsection{Method \code{optimiseMapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$optimiseMapping(
  alpha = 4,
  x0 = 0.5,
  nu = 1,
  optim.method = "BFGS",
  plot.3d = FALSE,
  dev.flat = NA,
  dev.polar = NA,
  control = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimiseMappingCart"></a>}}
\subsection{Method \code{optimiseMappingCart()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$optimiseMappingCart(
  alpha = 4,
  x0 = 0.5,
  nu = 1,
  method = "BFGS",
  plot.3d = FALSE,
  dev.flat = NA,
  dev.polar = NA,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transformImage"></a>}}
\subsection{Method \code{transformImage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$transformImage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIms"></a>}}
\subsection{Method \code{getIms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$getIms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTearCoords"></a>}}
\subsection{Method \code{getTearCoords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$getTearCoords(r)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeatureSet"></a>}}
\subsection{Method \code{getFeatureSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$getFeatureSet(type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reconstructFeatureSets"></a>}}
\subsection{Method \code{reconstructFeatureSets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$reconstructFeatureSets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPoints"></a>}}
\subsection{Method \code{getPoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$getPoints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapFlatToSpherical"></a>}}
\subsection{Method \code{mapFlatToSpherical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$mapFlatToSpherical(P)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedOutline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
