% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedPointSet.R
\name{ReconstructedPointSet}
\alias{ReconstructedPointSet}
\title{ReconstructedPointSet class}
\value{
An \code{ReconstructedPointSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
ReconstructedPointSet class

ReconstructedPointSet class
}
\author{
David Sterratt
}
\section{Super class}{
\code{\link[retistruct:ReconstructedFeatureSet]{retistruct::ReconstructedFeatureSet}} -> \code{ReconstructedPointSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getMean}{\code{ReconstructedPointSet$getMean()}}
\item \href{#method-getHullarea}{\code{ReconstructedPointSet$getHullarea()}}
\item \href{#method-getKDE}{\code{ReconstructedPointSet$getKDE()}}
\item \href{#method-clone}{\code{ReconstructedPointSet$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getCol">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getCol}{\code{retistruct::ReconstructedFeatureSet$getCol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getFeature">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getFeature}{\code{retistruct::ReconstructedFeatureSet$getFeature()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getID">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getID}{\code{retistruct::ReconstructedFeatureSet$getID()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getIDs">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getIDs}{\code{retistruct::ReconstructedFeatureSet$getIDs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="initialize">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-initialize}{\code{retistruct::ReconstructedFeatureSet$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMean"></a>}}
\subsection{Method \code{getMean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedPointSet$getMean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHullarea"></a>}}
\subsection{Method \code{getHullarea()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedPointSet$getHullarea()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getKDE"></a>}}
\subsection{Method \code{getKDE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedPointSet$getKDE()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedPointSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
