% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeTear}
\alias{removeTear}
\title{Remove tear from an AnnotatedOutline}
\usage{
removeTear(o, tid)
}
\arguments{
\item{o}{\code{AnnotatedOutline} object}

\item{tid}{Tear ID, which can be returned from \code{whichTear()}}
}
\value{
\code{AnnotatedOutline} object
}
\description{
Remove tear from an AnnotatedOutline
}
\author{
David Sterratt
}

