% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{report}
\alias{report}
\title{Reporting utility function}
\usage{
report(x, ...)
}
\arguments{
\item{x}{First arguments to reporting function}

\item{...}{Arguments to reporting function}
}
\description{
Calls function specified by option \code{retistruct.report}
}
\author{
David Sterratt
}
