% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct-batch.R
\name{retistruct.batch.analyse.summaries}
\alias{retistruct.batch.analyse.summaries}
\title{Extract statistics from a directory containing
reconstruction directories.}
\usage{
retistruct.batch.analyse.summaries(path)
}
\arguments{
\item{path}{Directory containing reconstruction directories}
}
\value{
Data frame containing various statistics
}
\description{
Extract statistics from a directory containing
reconstruction directories.
}
\author{
David Sterratt
}
