% Generated by roxygen2 (4.0.0): do not edit by hand
\name{retistruct.batch.analyse.summary}
\alias{retistruct.batch.analyse.summary}
\title{Extract statistics from the retistruct-batch.csv summary file}
\usage{
retistruct.batch.analyse.summary(path)
}
\arguments{
\item{path}{The path to the retistruct-batch.csv}
}
\value{
list of various statistics
}
\description{
Extract statistics from the retistruct-batch.csv summary file
}
\author{
David Sterratt
}

