\name{retistruct.batch.plot.ods}
\alias{retistruct.batch.plot.ods}
\title{Superposed plot of ODs on polar axes}
\usage{
retistruct.batch.plot.ods(summ, phi0d, ...)
}
\arguments{
  \item{summ}{Summary object returned by
  \code{\link{retistruct.batch.summary}}}

  \item{phi0d}{The rim angle for the plot}

  \item{...}{Other parameters, passed to projection}
}
\value{
A pseudo retina, in which the optic disks are treated as
datapoints
}
\description{
Polar plot of ODs of a group of retinae.
}
\author{
David Sterratt
}

