% Generated by roxygen2 (4.0.2): do not edit by hand
\name{retistruct.batch.plot.titrations}
\alias{retistruct.batch.plot.titrations}
\title{Plot titrations}
\usage{
retistruct.batch.plot.titrations(tdat)
}
\arguments{
\item{tdat}{Output of \code{\link{retistruct.batch.get.titrations}}}
}
\description{
Plot titrations
}

