% Generated by roxygen2 (4.0.0): do not edit by hand
\name{retistruct.batch.summary}
\alias{retistruct.batch.summary}
\title{Extract summary data for a batch of reconstructions}
\usage{
retistruct.batch.summary(tldir = ".", cache = TRUE)
}
\arguments{
\item{tldir}{The top level directory of the tree through which to
recurse.}

\item{cache}{If \code{TRUE} use the cached statistics rather than
generate on the fly (which is slower).}
}
\value{
Data frame containing summary data
}
\description{
Recurse through a directory tree, determining whether the
directory contains valid derived data and extracting summary data
if it does.
}
\author{
David Sterratt
}

