\name{retistruct.check.markup}
\alias{retistruct.check.markup}
\title{Retistruct check markup}
\usage{
retistruct.check.markup(o)
}
\arguments{
  \item{o}{Outline object}
}
\value{
If all markup is present, return \code{TRUE}. Otherwise
return \code{FALSE}.
}
\description{
Check that markup such as tears and the nasal or dorsal
points are present.
}
\author{
David Sterratt
}

