% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct-cli.R
\name{retistruct.cli.process}
\alias{retistruct.cli.process}
\title{Process a dataset, saving results to disk}
\usage{
retistruct.cli.process(dataset, outputdir = NA, device = "pdf")
}
\arguments{
\item{dataset}{Path to dataset to process}

\item{outputdir}{Directory in which to save any figures}

\item{device}{String representing device to print figures to}
}
\description{
This function processes a \code{dataset}, saving the
reconstruction data and MATLAB export data to the \code{dataset}
directory and printing figures to \code{outputdir}.
}
\author{
David Sterratt
}
