% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct.R
\name{retistruct.potential.od}
\alias{retistruct.potential.od}
\title{Test for a potential optic disc}
\usage{
retistruct.potential.od(o)
}
\arguments{
\item{o}{Outline object}
}
\value{
\code{TRUE} if an optic disc may be present; \code{FALSE} otherwise
}
\description{
Test the outline object \code{o} for the presence of
potential optic disc. This is done by checking that the list of
landmark lines \code{Ss} exists.
}
\author{
David Sterratt
}
