% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct.R
\name{retistruct.read.dataset}
\alias{retistruct.read.dataset}
\title{Read a retinal dataset}
\usage{
retistruct.read.dataset(dataset, report = message, ...)
}
\arguments{
\item{dataset}{Path to directory containing the files
corresponding to each format.}

\item{report}{Function to report progress. Set to \code{FALSE} for
no reporting.}

\item{...}{Parameters passed to the format-specific functions.}
}
\value{
A \code{\link{RetinalOutline}} object
}
\description{
Read a retinal dataset in one of three formats; for information on
formats see see \code{\link{idt.read.dataset}},
\code{\link{csv.read.dataset}} and
\code{\link{ijroi.read.dataset}}. The format is autodetected from
the files in the directory.
}
\author{
David Sterratt
}
