% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct.R
\name{retistruct.reconstruct}
\alias{retistruct.reconstruct}
\title{Reconstruct a retina}
\usage{
retistruct.reconstruct(
  a,
  report = NULL,
  plot.3d = FALSE,
  dev.flat = NA,
  dev.polar = NA,
  debug = FALSE,
  ...
)
}
\arguments{
\item{a}{\code{\link{RetinalOutline}} object with tear and
correspondence annotations}

\item{report}{Function to report progress. Set to \code{FALSE} for
no reporting or to \code{NULL} to inherit from the argument given to \code{\link{retistruct.read.dataset}}}

\item{plot.3d}{If \code{TRUE} show progress in a 3D plot}

\item{dev.flat}{The ID of the device to which to plot the flat
representation}

\item{dev.polar}{The ID of the device to which to plot the polar
representation}

\item{debug}{If \code{TRUE} print extra debugging output}

\item{...}{Parameters to be passed to
\code{\link{RetinalReconstructedOutline}} constructor}
}
\value{
A \code{\link{RetinalReconstructedOutline}} object
}
\description{
Reconstruct a retina
}
\author{
David Sterratt
}
