% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct.R
\name{retistruct.save.markup}
\alias{retistruct.save.markup}
\title{Save markup}
\usage{
retistruct.save.markup(a)
}
\arguments{
\item{a}{\code{\link{RetinalOutline}} object}
}
\description{
Save the markup in the \code{\link{RetinalOutline}} \code{a} to a
file called \code{markup.csv} in the directory \code{a$dataset}.
}
\author{
David Sterratt
}
