% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct.R
\name{retistruct.save.recdata}
\alias{retistruct.save.recdata}
\title{Save reconstruction data}
\usage{
retistruct.save.recdata(r)
}
\arguments{
\item{r}{\code{\link{RetinalReconstructedOutline}} object}
}
\description{
Save the reconstruction data in an object \code{r} of class
\code{\link{RetinalReconstructedOutline}} to a file called
\code{r.Rdata} in the directory \code{r$dataset}.
}
\author{
David Sterratt
}
