% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Retistruct Shiny Server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{object that holds the UI state (Managed automatically by shiny)}

\item{output}{sends new outputs to the UI (Managed automatically by shiny)}

\item{session}{controls each open instance (Managed automatically by shiny)}
}
\description{
The R shiny server responsible for storing a state for each
session, handling inputs from the UI to the server, and plotting outputs
to the UI. The arguments are all handled by the shiny package and this
function should not be instantiated manually.
}
\author{
Jan Okul
}
