% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Outline.R
\name{simplify.outline}
\alias{simplify.outline}
\title{Simplify an outline object by removing short edges}
\usage{
simplify.outline(o, min.frac.length = 0.001, plot = FALSE)
}
\arguments{
\item{o}{\code{outline} object to simplify}

\item{min.frac.length}{the minumum length as a fraction of the
total length of the outline.}

\item{plot}{whether to display plotting or not during simplification}
}
\value{
Simlified \code{outline} object
}
\description{
Simplify an outline object by removing verticies bordering short
edges while not encroaching on any of the outline. At present,
this is done by finding concave vertices. It is safe to remove
these, at the expense of increasing the area a bit.
}
\author{
David Sterratt
}

