% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geometry.R
\name{sphere.cart.to.sphere.wedge}
\alias{sphere.cart.to.sphere.wedge}
\title{Convert from Cartesian to 'wedge' coordinates}
\usage{
sphere.cart.to.sphere.wedge(P, phi0, R = 1)
}
\arguments{
\item{P}{locations of points on sphere as N-by-3 matrix with
labelled columns "X", "Y" and "Z"}

\item{phi0}{rim angle as colatitude}

\item{R}{radius of sphere}
}
\value{
2-column Matrix of 'wedge' coordinates of points on
sphere. Column names are \code{phi} and \code{lambda}.
}
\description{
Convert points in 3D cartesian space to locations of points on
sphere in 'wedge' coordinates (\var{psi}, \var{f}).  Wedges are
defined by planes inclined at an angle \var{psi} running through a
line between poles on the rim above the x axis.  \var{f} is the
fractional distance along the circle defined by the intersection
of this plane and the curtailed sphere.
}
\author{
David Sterratt
}

