% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometry.R
\name{sphere.spherical.to.polar.cart}
\alias{sphere.spherical.to.polar.cart}
\title{Convert spherical coordinates on sphere to  polar
projection in Cartesian coordinates}
\usage{
sphere.spherical.to.polar.cart(r, pa = FALSE, preserve = "lattitude")
}
\arguments{
\item{r}{2-column Matrix of spherical coordinates of points on
sphere. Column names are \code{phi} and \code{lambda}.}

\item{pa}{If \code{TRUE}, make this an area-preserving projection}

\item{preserve}{Quantity to preserve locally in the
projection. Options are \code{lattitude}, \code{area} or
\code{angle}}
}
\value{
2-column Matrix of Cartesian coordinates of points on polar
projection. Column names should be \code{x} and \code{y}
}
\description{
This is the inverse of \code{\link{polar.cart.to.sphere.spherical}}
}
\author{
David Sterratt
}

