% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sphericalplot}
\alias{sphericalplot}
\title{Spherical plot of reconstructed outline}
\usage{
sphericalplot(r, ...)
}
\arguments{
\item{r}{Object inheriting \code{\link{ReconstructedOutline}}}

\item{...}{Parameters depending on class of \code{r}}
}
\description{
Spherical plot of reconstructed outline
}
\author{
David Sterratt
}
