\name{sphericalplot.reconstructedOutline}
\alias{sphericalplot.reconstructedOutline}
\title{Spherical plot of reconstructed outline}
\usage{
  \method{sphericalplot}{reconstructedOutline} (r,
    strain = FALSE, surf = TRUE, ...)
}
\arguments{
  \item{r}{\code{reconstructedOutline} object}

  \item{strain}{If \code{TRUE}, plot the strain}

  \item{surf}{If \code{TRUE}, plot the surface}

  \item{...}{Other graphics parameters -- not used at
  present}
}
\description{
  Draw a spherical plot of reconstructed outline. This
  method just draws the mesh.
}
\author{
  David Sterratt
}

