\name{strain.colours}
\alias{strain.colours}
\title{Generate colours for strain plots}
\usage{
strain.colours(x)
}
\arguments{
  \item{x}{Vector of values of log strain}
}
\value{
Vector of colours corresponding to strains
}
\description{
Generate colours for strain plots
}
\author{
David Sterratt
}

