\name{trace.ray.angle}
\alias{trace.ray.angle}
\title{Trace ray at an angle}
\usage{
  trace.ray.angle(psi, anp, S)
}
\arguments{
  \item{psi}{Angle from axis}

  \item{anp}{Nodal point (hopefully anterior nodal point)}

  \item{S}{System}
}
\value{
  Ray - see \code{\link{trace.ray}}
}
\description{
  Trace ray at an angle
}
\author{
  David Sterratt
}

