% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tri.area}
\alias{tri.area}
\title{Area of triangles on a plane}
\usage{
tri.area(P, Pt)
}
\arguments{
\item{P}{2-column matrix of vertices of triangles}

\item{Pt}{3-column matrix of indicies of rows of \code{P} giving
triangulation}
}
\value{
Vectors of areas of triangles
}
\description{
Area of triangles on a plane
}
\author{
David Sterratt
}

