% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangulatedFragment.R
\name{TriangulatedFragment}
\alias{TriangulatedFragment}
\title{Triangulate outline

Parameters to constructor:
\code{P} Points to triangulate
\code{n} Minimum number of points in the triangulation
\code{suppress.external.steiner} If \code{TRUE} prevent the
addition of points in the outline. This happens to maintain
triangle quality.
\code{report} Function to report progress}
\value{
A \code{triangulatedFragment} object containing the
following fields:
\item{\code{P}}{The set of new points, with the existing points at the start}
\item{\code{T}}{The triangulation}
\item{\code{Cu}}{Unique set of M connections, as M*2 matrix}
\item{\code{h}}{Correspondences mapping}
\item{\code{A}}{Array containing area of each triangle}
\item{\code{L}}{Length of each connection}
\item{\code{A.signed}}{Signed area of each triangle}
\item{\code{A.tot}}{Total area of outline}
\item{\code{gf}}{Forward pointers}
\item{\code{gb}}{Backward pointers}
\item{\code{S}}{Segments (from \code{\link{triangulate}})}
\item{\code{E}}{Edges (from \code{\link{triangulate}})}
\item{\code{EB}}{Edge boundaries (from \code{\link{triangulate}})}
}
\description{
Create a triangulation of the \code{Fragment} object \code{o}.  The
minimum number of triangles in the triangulation is specified by
\code{n}.
}
\author{
David Sterratt
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TriangulatedFragment$new()}}
\item \href{#method-clone}{\code{TriangulatedFragment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TriangulatedFragment$new(
  fragment,
  n = 200,
  suppress.external.steiner = FALSE,
  report = message
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TriangulatedFragment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
