% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Retistruct UI}
\usage{
ui()
}
\description{
The Shiny UI element, runs on a browser and is similar to HTML, attempted to
mimic the original Retistruct UI as closely as possible.
}
\author{
Jan Okul
}
