\name{whichTear}
\alias{whichTear}
\title{Return index of tear}
\usage{
  whichTear(o, pid)
}
\arguments{
  \item{o}{\code{AnnotatedOutline} object}

  \item{pid}{ID of point}
}
\value{
  ID of tear
}
\description{
  Return index of tear in an AnnotatedOutline in which a
  point appears
}
\author{
  David Sterratt
}

