\name{addTear}
\alias{addTear}
\title{Add tear to an AnnotatedOutline}
\usage{
addTear(a, pids)
}
\arguments{
  \item{a}{\code{AnnotatedOutline} object}

  \item{pids}{Vector of three point IDs to be added}
}
\value{
\code{AnnotatedOutline} object
}
\description{
Add tear to an AnnotatedOutline
}
\author{
David Sterratt
}

