% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\docType{data}
\name{AnnotatedOutline}
\alias{AnnotatedOutline}
\title{AnnotatedOutline object}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
AnnotatedOutline
}
\value{
AnnotatedOutline object, with extra fields for tears
  latitude of rim \code{phi0} and index of fixed point \code{i0}.
}
\description{
Constructor for AnnotatedOutline object.
}
\examples{
P <- rbind(c(1,1),   c(2,1),  c(2,-1),
           c(1,-1),  c(1,-2), c(-1,-2),
           c(-1,-1), c(-2,-1),c(-2,1),
           c(-1,1),  c(-1,2), c(1,2))
o <- TriangulatedOutline$new(P)
o$addTear(c(3, 4, 5))
o$addTear(c(6, 7, 8))
o$addTear(c(9, 10, 11))
o$addTear(c(12, 1, 2))
flatplot(o)
}
\author{
David Sterratt
}
\keyword{datasets}
