% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\name{checkTears}
\alias{checkTears}
\title{Check that tears are all in the correct direction}
\usage{
checkTears(o)
}
\arguments{
\item{o}{\code{AnnotatedOutline} object}
}
\value{
If all is OK, returns empty vector.  If not, returns
indicies of problematic tears.
}
\description{
Given a tear matrix T with columns "V0", "VF", and "VB", check
that all tears are correct.
}
\author{
David Sterratt
}

