% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-csv.R
\name{csv.read.dataset}
\alias{csv.read.dataset}
\title{Read a retinal dataset in CSV format}
\usage{
csv.read.dataset(dataset, report = message)
}
\arguments{
\item{dataset}{Path to directory containing \code{outline.csv}}

\item{report}{Function to report progress}
}
\value{
A \code{\link{RetinalOutline}} object
}
\description{
Read a retinal dataset in CSV format. Each dataset is a folder
containing a file called outline.csv that specifies the outline in
X-Y coordinates. It may also contain a file \code{datapoints.csv},
containing the locations of data points and a file
\code{datacounts.csv}, containing the locations of data counts;
see \code{\link{read.datapoints}} and
\code{\link{read.datacounts}} for the formats of these files. The
folder may also contain a file \code{od.csv} specifying the
coordinates of the optic disc.
}
\author{
David Sterratt
}
