% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format-csv.R
\name{csv.read.dataset}
\alias{csv.read.dataset}
\title{Read a retinal dataset in CSV format}
\usage{
csv.read.dataset(dataset)
}
\arguments{
\item{dataset}{Path to directory containing \code{outline.csv}}
}
\value{
A \code{\link{RetinalDataset}} object
}
\description{
Read a retinal dataset in CSV format. Each dataset is a folder
containing a file called outline.csv that specifies the outline in
X-Y coordinates. It may also contain a file \code{datapoints.csv},
containing the locations of data points; see
\code{\link{read.datapoints}} for the format of this file. The
folder may also contain a file \code{od.csv} specifying the
coordinates of the optic disc.
}
\author{
David Sterratt
}

