% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSet.R
\docType{data}
\name{FeatureSet}
\alias{FeatureSet}
\title{FeatureSet class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FeatureSet
}
\value{
An \code{FeatureSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
FeatureSet class
}
\author{
David Sterratt
}
\keyword{datasets}
