% Generated by roxygen2 (4.0.0): do not edit by hand
\name{flatplot.annotatedOutline}
\alias{flatplot.annotatedOutline}
\title{Flat plot of AnnotatedOutline}
\usage{
\method{flatplot}{annotatedOutline}(x, axt = "n", ylim = NULL,
  markup = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{AnnotatedOutline}} object}

\item{axt}{whether to plot axes}

\item{ylim}{y-limits}

\item{markup}{If \code{TRUE}, plot markup}

\item{...}{Other plotting parameters}
}
\description{
Plot flat \code{\link{AnnotatedOutline}}. The user markup is
displayed by default.
}
\author{
David Sterratt
}

