% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{flatplot}
\alias{flatplot}
\title{Plot "flat" (unreconstructed) representation of outline}
\usage{
flatplot(x, axt = "n", xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{\code{\link{Outline}}, \code{\link{AnnotatedOutline}},  \code{\link{StitchedOutline}} &c object}

\item{axt}{whether to plot axes}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{...}{Other plotting parameters}
}
\description{
Plot "flat" (unreconstructed) representation of outline
}
\author{
David Sterratt
}
