\name{flatplot}
\alias{flatplot}
\title{Flat plot of object}
\usage{
  flatplot(x, axt = "n", ylim = NULL, ...)
}
\arguments{
  \item{x}{\code{\link{Outline}}, \code{\link{Dataset}} \&c
  object}

  \item{axt}{whether to plot axes}

  \item{ylim}{y-limits}

  \item{...}{Other plotting parameters}
}
\description{
  Plot flat representation of object
}
\author{
  David Sterratt
}

