% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StitchedOutline.R
\name{flatplot.stitchedOutline}
\alias{flatplot.stitchedOutline}
\title{Flat plot of AnnotatedOutline}
\usage{
\method{flatplot}{stitchedOutline}(x, axt = "n", ylim = NULL,
  stitch = TRUE, lwd = 1, ...)
}
\arguments{
\item{x}{\code{\link{AnnotatedOutline}} object}

\item{axt}{whether to plot axes}

\item{ylim}{y-limits}

\item{stitch}{If \code{TRUE}, plot stitch}

\item{lwd}{Line width}

\item{...}{Other parameters}
}
\description{
Plot flat \code{\link{StitchedOutline}}. If the optional argument
\code{stitch} is \code{TRUE} the user markup is displayed.
}
\author{
David Sterratt
}

