% Generated by roxygen2 (4.0.0): do not edit by hand
\name{flatplot.triangulatedOutline}
\alias{flatplot.triangulatedOutline}
\title{Flat plot of TriangulatedOutline}
\usage{
\method{flatplot}{triangulatedOutline}(x, axt = "n", ylim = NULL,
  mesh = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{TriangulatedOutline}} object}

\item{axt}{whether to plot axes}

\item{ylim}{y-limits}

\item{mesh}{If \code{TRUE}, plot mesh}

\item{...}{Other plotting parameters}
}
\description{
Plot flat \code{\link{TriangulatedOutline}}.
}
\author{
David Sterratt
}

