\name{fp}
\alias{fp}
\title{Piecewise smooth function used in area penalty}
\usage{
fp(x, x0)
}
\arguments{
  \item{x}{Main argument}

  \item{x0}{The cutoff parameter. Above this value the
  function is zero.}
}
\value{
The value of the function.
}
\description{
Derivative of \code{\link{f}}
}
\author{
David Sterratt
}

