\name{getDss}
\alias{getDss}
\title{Get transformed spherical coordinates of datapoints}
\usage{
  getDss(r)
}
\arguments{
  \item{r}{\code{\link{ReconstructedDataset}} or
  \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
  \code{Dss}
}
\description{
  Get spherical coordinates of datapoints.
}
\author{
  David Sterratt
}

