\name{getDss.retinalReconstructedDataset}
\alias{getDss.retinalReconstructedDataset}
\title{Get transformed spherical coordinates of datapoints}
\usage{
\method{getDss}{retinalReconstructedDataset}(r)
}
\arguments{
  \item{r}{\code{\link{RetinalReconstructedDataset}}
  object.}
}
\value{
\code{Dss}
}
\description{
Get spherical coordinates of datapoints, transformed
according to the values of \code{DVflip} and \code{side}.
}
\author{
David Sterratt
}

