\name{getDssMean}
\alias{getDssMean}
\title{Karcher mean of datapoints in spherical coordinates}
\usage{
getDssMean(r)
}
\arguments{
  \item{r}{\code{\link{ReconstructedDataset}} or
  \code{\link{RetinalReconstructedDataset}} object.}
}
\value{
\code{Dss.mean}
}
\description{
Get Karcher mean of datapoints in spherical coordinates.
}
\author{
David Sterratt
}

