\name{getKDE}
\alias{getKDE}
\title{Get kernel density estimate of data points}
\usage{
  getKDE(r)
}
\arguments{
  \item{r}{\code{\link{ReconstructedDataset}} object}
}
\value{
  See \code{\link{compute.kernel.estimate}}
}
\description{
  Get kernel density estimate of data points
}
\author{
  David Sterratt
}

