% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReconstructedDataset.R
\name{getKR}
\alias{getKR}
\title{Get kernel regression estimate of grouped data points}
\usage{
getKR(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} object}
}
\value{
See \code{\link{compute.kernel.estimate}}
}
\description{
Get kernel regression estimate of grouped data points
}
\author{
David Sterratt
}

