% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getKR}
\alias{getKR}
\title{Get kernel regression estimate of grouped data points}
\usage{
getKR(r)
}
\arguments{
\item{r}{\code{\link{ReconstructedDataset}} object}
}
\value{
See \code{\link{compute.kernel.estimate}}
}
\description{
Get kernel regression estimate of grouped data points
}
\author{
David Sterratt
}

