\name{getSss.reconstructedDataset}
\alias{getSss.reconstructedDataset}
\title{Get transformed spherical coordinates of landmarks.}
\usage{
\method{getSss}{reconstructedDataset}(r)
}
\arguments{
  \item{r}{\code{ReconstructedDataset} object.}
}
\value{
\code{Sss}
}
\description{
Get spherical coordinates of landmarks.
}
\author{
David Sterratt
}

