\name{getTear}
\alias{getTear}
\title{Return indicies of tear in AnnotatedOutline}
\usage{
getTear(o, tid)
}
\arguments{
  \item{o}{\code{AnnotatedOutline} object}

  \item{tid}{Tear ID, which can be returned from
  \code{whichTear()}}
}
\value{
Vector of three point IDs, labelled with \code{V0},
\code{VF} and \code{VB}
}
\description{
Return indicies of tear in AnnotatedOutline
}
\author{
David Sterratt
}

