\name{kde.fhat.cart}
\alias{kde.fhat.cart}
\title{Kernel density estimate on sphere using Fisherian density
with Cartesian coordinates}
\usage{
kde.fhat.cart(r, mu, kappa)
}
\arguments{
  \item{r}{Locations at which to estimate density in
  Cartesian coordinates on unit sphere}

  \item{mu}{Locations of data points in Cartesian
  coordinates on unit sphere}

  \item{kappa}{Concentration parameter}
}
\value{
Vector of density estimates
}
\description{
Kernel density estimate on sphere using Fisherian density
with Cartesian coordinates
}
\author{
David Sterratt
}

