% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel-density.R
\name{kr.compute.concentration}
\alias{kr.compute.concentration}
\title{Find the optimal concentration for a set of data}
\usage{
kr.compute.concentration(mu, y)
}
\arguments{
\item{mu}{Locations in Cartesian coordinates (independent variables)}

\item{y}{Values at locations (dependent variables)}
}
\value{
The optimal concentration
}
\description{
Find the optimal concentration for a set of data
}
\author{
David Sterratt
}

