% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\name{labelTearPoints}
\alias{labelTearPoints}
\title{Label three outline point indicies as apicies and vertices of tear}
\usage{
labelTearPoints(o, m)
}
\arguments{
\item{o}{Outline object}

\item{m}{the vector of three indicies}
}
\value{
Vector of indicies labelled with V0, VF and VB
}
\description{
Label a set of three unlabelled points supposed to refer to the
apex and vertcies of a cut and tear with the V0 (Apex), VF
(forward vertex) and VB (backward vertex) labels.
}
\author{
David Sterratt
}

