% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandmarkSet.R
\name{LandmarkSet}
\alias{LandmarkSet}
\title{Subclass of \code{\link{FeatureSet}} to represent points}
\description{
A LandmarkSet contains information about points
  located on \code{\link{Outline}}s. Each LandmarkSet contains a
  list of matrices, each of which has columns labelled \code{X}
  and \code{Y} describing the cartesian coordinates (in the
  unscaled coordinate frame) of points in landmarks on the
  Outline.
}
\author{
David Sterratt
}
\section{Super classes}{
\code{\link[retistruct:FeatureSetCommon]{retistruct::FeatureSetCommon}} -> \code{\link[retistruct:FeatureSet]{retistruct::FeatureSet}} -> \code{LandmarkSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LandmarkSet-new}{\code{LandmarkSet$new()}}
\item \href{#method-LandmarkSet-reconstruct}{\code{LandmarkSet$reconstruct()}}
\item \href{#method-LandmarkSet-clone}{\code{LandmarkSet$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSetCommon" data-id="getCol"><a href='../../retistruct/html/FeatureSetCommon.html#method-FeatureSetCommon-getCol'><code>retistruct::FeatureSetCommon$getCol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSetCommon" data-id="getFeature"><a href='../../retistruct/html/FeatureSetCommon.html#method-FeatureSetCommon-getFeature'><code>retistruct::FeatureSetCommon$getFeature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSetCommon" data-id="getFeatures"><a href='../../retistruct/html/FeatureSetCommon.html#method-FeatureSetCommon-getFeatures'><code>retistruct::FeatureSetCommon$getFeatures()</code></a></span></li>
<li><span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSetCommon" data-id="getIDs"><a href='../../retistruct/html/FeatureSetCommon.html#method-FeatureSetCommon-getIDs'><code>retistruct::FeatureSetCommon$getIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSetCommon" data-id="getIndex"><a href='../../retistruct/html/FeatureSetCommon.html#method-FeatureSetCommon-getIndex'><code>retistruct::FeatureSetCommon$getIndex()</code></a></span></li>
<li><span class="pkg-link" data-pkg="retistruct" data-topic="FeatureSetCommon" data-id="setID"><a href='../../retistruct/html/FeatureSetCommon.html#method-FeatureSetCommon-setID'><code>retistruct::FeatureSetCommon$setID()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandmarkSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-LandmarkSet-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandmarkSet$new(data = NULL, cols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{List of matrices describing data. Each matrix
should have columns named \code{X} and \code{Y}}

\item{\code{cols}}{Named vector of colours for each data set. The name is
used as the ID (label) for the data set. The colours should be names
present in the output of the \code{\link{colors}} function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandmarkSet-reconstruct"></a>}}
\if{latex}{\out{\hypertarget{method-LandmarkSet-reconstruct}{}}}
\subsection{Method \code{reconstruct()}}{
Map the LandmarkSet to a \code{\link{ReconstructedOutline}}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandmarkSet$reconstruct(ro)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ro}}{The \code{\link{ReconstructedOutline}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandmarkSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LandmarkSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandmarkSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
