% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outline.R
\docType{data}
\name{Outline}
\alias{Outline}
\title{Outline class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Outline
}
\value{
An \code{Outline} object containing the following:
\item{\code{P}}{A N-by-2 matrix of points of the \code{Outline} arranged in anticlockwise order}
\item{\code{gf}}{For each row of \code{P}, the index of \code{P} that is next in the outline travelling anticlockwise (forwards)}
\item{\code{gb}}{For each row of \code{P}, the index of \code{P} that is next in the outline travelling clockwise (backwards)}
\item{\code{h}}{For each row of \code{P}, the correspondence of that point (which will be to itself initially)}
\item{\code{im}}{The image as a \code{raster} object}
\item{\code{scale}}{The length of one unit of \code{P} in arbitrary units}
}
\description{
Construct an outline object. This sanitises the input points
\code{P}, as described below.
}
\author{
David Sterratt
}
\keyword{datasets}
