% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{panlabel}
\alias{panlabel}
\title{Ancillary function to place labels}
\usage{
panlabel(panlabel, line = -0.7)
}
\arguments{
\item{panlabel}{Label text}

\item{line}{Line on which to appear}
}
\description{
Ancillary function to place labels
}
\author{
David Sterratt
}

