% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathOutline.R
\docType{data}
\name{PathOutline}
\alias{PathOutline}
\title{Add point correspondences to the outline}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
PathOutline
}
\value{
To the \code{\link{Outline}} object this adds
\item{\code{hf}}{point correspondence mapping in forward direction for
points on boundary}
\item{\code{hb}}{point correspondence mapping in backward direction for
points on boundary}
}
\description{
The member function \code{stitchSubpaths()} stitches together two
subpaths of the outline. One subpath is stitched in the forward
direction from the point indexed by \code{VF0} to the point
indexed by \code{VF1}. The other is stitched in the backward
direction from \code{VB0} to \code{VB1}. Each point in the subpath
is linked to points in the opposing pathway at an equal or
near-equal fraction along. If a point exists in the opposing
pathway within a distance \code{epsilon} of the projection, this
point is connected. If no point exists within this tolerance, a
new point is created.
}
\keyword{datasets}
