% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSet.R
\docType{data}
\name{PointSet}
\alias{PointSet}
\title{PointSet class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
PointSet
}
\value{
An \code{PointSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
PointSet class
}
\author{
David Sterratt
}
\keyword{datasets}
