\name{projection}
\alias{projection}
\title{Plot projection of an object}
\usage{
projection(r, ...)
}
\arguments{
  \item{r}{Object such as a
  \code{\link{ReconstructedOutline}}}

  \item{...}{Other parameters; see
  \code{\link{projection.reconstructedOutline}} and
  \code{\link{projection.reconstructedDataset}}}
}
\description{
Generic function for plotting projections of objects.
}
\author{
David Sterratt
}

