% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{recfile.version}
\alias{recfile.version}
\title{Version of reconstruction file data format}
\format{\preformatted{ num 5
}}
\usage{
recfile.version
}
\description{
Version of reconstruction file data format
}
\keyword{datasets}

