% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retistruct.R
\docType{data}
\name{recfile.version}
\alias{recfile.version}
\title{Version of reconstruction file data format}
\format{\preformatted{ num 5
}}
\usage{
recfile.version
}
\description{
Version of reconstruction file data format
}
\keyword{datasets}

