% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedFeatureSet.R
\docType{data}
\name{ReconstructedFeatureSet}
\alias{ReconstructedFeatureSet}
\title{ReconstructedFeatureSet class
Initialised with a FeatureSet (\code{fs}) and an ReconstructedOutline (\code{ro}) object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReconstructedFeatureSet
}
\value{
A \code{ReconstructedFeatureSet} object.
}
\description{
ReconstructedFeatureSet class
Initialised with a FeatureSet (\code{fs}) and an ReconstructedOutline (\code{ro}) object
}
\author{
David Sterratt
}
\keyword{datasets}
