% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedLandmarkSet.R
\name{ReconstructedLandmarkSet}
\alias{ReconstructedLandmarkSet}
\title{ReconstructedLandmarkSet class}
\value{
An \code{ReconstructedLandmarkSet} object, which inherits
  from ReconstructedFeatureSet
}
\description{
ReconstructedLandmarkSet class

ReconstructedLandmarkSet class
}
\author{
David Sterratt
}
\section{Super class}{
\code{\link[retistruct:ReconstructedFeatureSet]{retistruct::ReconstructedFeatureSet}} -> \code{ReconstructedLandmarkSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{ReconstructedLandmarkSet$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getCol">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getCol}{\code{retistruct::ReconstructedFeatureSet$getCol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getFeature">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getFeature}{\code{retistruct::ReconstructedFeatureSet$getFeature()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getID">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getID}{\code{retistruct::ReconstructedFeatureSet$getID()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="getIDs">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-getIDs}{\code{retistruct::ReconstructedFeatureSet$getIDs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="retistruct" data-topic="ReconstructedFeatureSet" data-id="initialize">}\href{../../retistruct/html/ReconstructedFeatureSet.html#method-initialize}{\code{retistruct::ReconstructedFeatureSet$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReconstructedLandmarkSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
