% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedLandmarkSet.R
\docType{data}
\name{ReconstructedLandmarkSet}
\alias{ReconstructedLandmarkSet}
\title{ReconstructedLandmarkSet class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReconstructedLandmarkSet
}
\value{
An \code{ReconstructedLandmarkSet} object, which inherits
  from ReconstructedFeatureSet
}
\description{
ReconstructedLandmarkSet class
}
\author{
David Sterratt
}
\keyword{datasets}
