% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructedPointSet.R
\docType{data}
\name{ReconstructedPointSet}
\alias{ReconstructedPointSet}
\title{ReconstructedPointSet class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
ReconstructedPointSet
}
\value{
An \code{ReconstructedPointSet} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
ReconstructedPointSet class
}
\author{
David Sterratt
}
\keyword{datasets}
