% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetinalOutline.R
\docType{data}
\name{RetinalOutline}
\alias{RetinalOutline}
\title{RetinalOutline class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RetinalOutline
}
\value{
An \code{retinalOutline} object. This contains the following fields:
\item{\code{DVflip}}{\code{TRUE} if the raw data is flipped in
the dorsoventral direction} 
\item{\code{side}}{The side of the eye ("Left" or "Right")}
\item{\code{dataset}}{File system path to dataset}
}
\description{
RetinalOutline class
}
\author{
David Sterratt
}
\keyword{datasets}
