% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetinalReconstructedOutline.R
\docType{data}
\name{RetinalReconstructedOutline}
\alias{RetinalReconstructedOutline}
\title{RetinalReconstructedOutline constructor}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
RetinalReconstructedOutline
}
\value{
\code{RetinalReconstructedOutline} object. This does not
contain any extra fields, but there are extra methods that apply
to it.
}
\description{
Create an object that is specific to retinal datasets. This
contains methods that return data point and landmark coordinates
that have been transformed according to the values of
\code{DVflip} and \code{side}.
}
\author{
David Sterratt
}
\keyword{datasets}
