\name{retistruct.batch.analyse.summaries}
\alias{retistruct.batch.analyse.summaries}
\title{Extract statistics from a directory containing
reconstruction directories.}
\usage{
retistruct.batch.analyse.summaries(path)
}
\arguments{
  \item{path}{Directory containing recontstruction
  directories}
}
\value{
Data frame containg various statistics
}
\description{
Extract statistics from a directory containing
reconstruction directories.
}
\author{
David Sterratt
}

