% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retistruct-batch.R
\name{retistruct.batch.analyse.summary}
\alias{retistruct.batch.analyse.summary}
\title{Extract statistics from the retistruct-batch.csv summary file}
\usage{
retistruct.batch.analyse.summary(path)
}
\arguments{
\item{path}{The path to the retistruct-batch.csv}
}
\value{
list of various statistics
}
\description{
Extract statistics from the retistruct-batch.csv summary file
}
\author{
David Sterratt
}

