% Generated by roxygen2 (4.0.0): do not edit by hand
\name{retistruct.cli.process}
\alias{retistruct.cli.process}
\title{Process a dataset, saving results to disk}
\usage{
retistruct.cli.process(dataset, outputdir = NA, device = "pdf",
  titrate = FALSE)
}
\arguments{
\item{dataset}{Path to dataset to process}

\item{outputdir}{Directory in which to save any figures}

\item{device}{String representing device to print figures to}

\item{titrate}{If \code{TRUE} add output of
\code{\link{titrate.reconstructedOutline}}  to object saved.}
}
\description{
This function processes a \code{dataset}, saving the
reconstruction data and matlab export data to the \code{dataset}
directory and printing figures to \code{outputdir}.
}
\author{
David Sterratt
}

