% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{retistruct.global.revision}
\alias{retistruct.global.revision}
\title{Git short commit hash of Retistruct package}
\format{\preformatted{ chr "72c7c84"
}}
\usage{
retistruct.global.revision
}
\description{
Git short commit hash of Retistruct package
}
\keyword{datasets}

