% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/revision.R
\docType{data}
\name{retistruct.global.revision}
\alias{retistruct.global.revision}
\title{Git short commit hash of Retistruct package}
\format{\preformatted{ chr "d573ca8"}}
\usage{
retistruct.global.revision
}
\description{
Git short commit hash of Retistruct package
}
\keyword{datasets}

