% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{retistruct.global.revision}
\alias{retistruct.global.revision}
\title{SVN global revision number of Retistruct package}
\format{\preformatted{ num 1316
}}
\usage{
retistruct.global.revision
}
\description{
SVN global revision number of Retistruct package
}
\keyword{datasets}

