% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retistruct-gui.R
\name{retistruct}
\alias{retistruct}
\title{Start the Retistruct GUI}
\usage{
retistruct()
}
\value{
Object with \code{getData()} method to return
reconstructed retina data and environment \code{this} which
contains variables in object.
}
\description{
Start the Retistruct GUI
}
\seealso{
gWidgets
}

