\name{retistruct.save.markup}
\alias{retistruct.save.markup}
\title{Save markup}
\usage{
  retistruct.save.markup(a)
}
\arguments{
  \item{a}{\code{\link{RetinalDataset}} object}
}
\description{
  Save the makrup in the \code{\link{RetinalDataset}}
  \code{a} to a file called \code{markup.csv} in the
  directory \code{a$dataset}.
}
\author{
  David Sterratt
}

