\name{retistruct.save.recdata}
\alias{retistruct.save.recdata}
\title{Save reconstruction data}
\usage{
retistruct.save.recdata(r)
}
\arguments{
  \item{r}{\code{\link{ReconstructedDataset}} object}
}
\description{
Save the reconstruction data in an object \code{r} that
inherits \code{\link{ReconstructedDataset}} and
\code{\link{ReconstructedOutline}} to a file called
\code{r.Rdata} in the directory \code{r$dataset}.
}
\author{
David Sterratt
}

