\name{sphericalplot}
\alias{sphericalplot}
\title{Spherical plot of reconstructed outline}
\usage{
sphericalplot(r, ...)
}
\arguments{
  \item{r}{Object with class
  \code{\link{ReconstructedOutline}} and/or
  \code{\link{ReconstructedDataset}}}

  \item{...}{Parameters depending on class of \code{r}}
}
\description{
Draw a spherical plot.
}
\author{
David Sterratt
}

