\name{stretchMesh}
\alias{stretchMesh}
\title{Stretch mesh}
\usage{
stretchMesh(Cu, L, i.fix, P.fix)
}
\arguments{
  \item{Cu}{Edge matrix}

  \item{L}{Lengths in flat outline}

  \item{i.fix}{Indicies of fixed points}

  \item{P.fix}{Coordinates of fixed points}
}
\value{
New matrix of 2D point locations
}
\description{
Stretch the mesh in the flat retina to a circular outline
}
\author{
David Sterratt
}

