% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TriangulatedOutline.R
\name{TriangulatedOutline}
\alias{TriangulatedOutline}
\title{Triangulate outline}
\usage{
TriangulatedOutline(o, n = 200, suppress.external.steiner = FALSE)
}
\arguments{
\item{o}{\code{\link{Outline}} object}

\item{n}{Minimum number of points in the triangulation}

\item{suppress.external.steiner}{If \code{TRUE} prevent the
addition of points in the outline. This happens to maintain
triangle quality.}
}
\value{
A \code{triangulatedOutline} object containing the
following fields:
\item{\code{P}}{The set of new points, with the existing points at the start}
\item{\code{T}}{The triangulation}
\item{\code{Cu}}{Unique set of M connections, as M*2 matrix}
\item{\code{h}}{Correspondances mapping}
\item{\code{A}}{Array containing area of each triangle}
\item{\code{L}}{Length of each connection}
\item{\code{A.signed}}{Signed area of each triangle}
\item{\code{A.tot}}{Total area of outline}
\item{\code{gf}}{Forward pointers}
\item{\code{gb}}{Backward pointers}
\item{\code{S}}{Segments (from \code{\link{triangulate}})}
\item{\code{E}}{Edges (from \code{\link{triangulate}})}
\item{\code{EB}}{Edge boundaries (from \code{\link{triangulate}})}
}
\description{
Create a triangulation of the \code{Outline} object \code{o}.  The
minimum number of triangles in the triangulation is specified by
\code{n}.
}
\author{
David Sterratt
}

