% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Retistruct UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
The Shiny UI element, runs on a browser and is similar to HTML, attempted to
mimic the original Retistruct UI as closely as possible.
}
\author{
Jan Okul
}
\keyword{datasets}
