% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotatedOutline.R
\name{whichTear}
\alias{whichTear}
\title{Return index of tear}
\usage{
whichTear(o, pid)
}
\arguments{
\item{o}{\code{AnnotatedOutline} object}

\item{pid}{ID of point}
}
\value{
ID of tear
}
\description{
Return index of tear in an AnnotatedOutline in which a point
appears
}
\author{
David Sterratt
}

